+++ src/CosmoFuse/__init__.py
@@ -25,6 +25,7 @@ from .multi_device import MultiDeviceCorrelation
 from .pipeline import MapFileLoader, MapLoader
 from .treecode import DEFAULT_RESOLUTION_FACTOR
 from .utils import pixel2RaDec, select_patch_centers
+from .zeta_writer import ZetaWriter
 
 __version__ = "6.0.0"
 __author__ = "David Gebauer"
@@ -51,4 +52,5 @@ __all__ = [
     "zeta_a_t",
     "pixel2RaDec",
     "select_patch_centers",
+    "ZetaWriter",
 ]